requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-allauth==0.21.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
flake8==2.4.1
gunicorn==19.3.0
mccabe==0.3.1
oauthlib==0.7.2
pep8==1.5.7
Pillow==2.9.0
psycopg2==2.6.1
pyflakes==0.8.1
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
static3==0.6.1
twython==3.2.0
wikipedia==1.4.0
